requirements.txt
apscheduler
python-dotenv
yfinance
requests
pandas